feat: add close list form on escape key
