+++ src/components/ListTitle.tsx
@@ -1,22 +1,17 @@
 import React, { useState } from 'react';
 import {
-  Typography, TextField, Tooltip, IconButton, Box,
+  Typography, Tooltip, IconButton, Box,
 } from '@mui/material';
-import CloseSharpIcon from '@mui/icons-material/CloseSharp';
 import DeleteIcon from '@mui/icons-material/Delete';
-import DoneSharpIcon from '@mui/icons-material/DoneSharp';
-import { useForm, Controller } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
 import { Column } from '../types/columns';
-import { editColumn, setColumn } from '../store/columnSlice';
-import { REMOVE_COLUMN, EDIT_COLUMN_TITLE } from '../constants/formfields';
+import { setColumn } from '../store/columnSlice';
+import { REMOVE_COLUMN } from '../constants/formfields';
 import { openModal } from '../store/modalSlice';
-import { DEFAULT_COLUMN } from '../constants/columns';
 import { AppDispatch } from '../store/store';
-import convertRulesRegExp from '../helpers/convertRulesRegExp';
-import { FormField } from '../types/formTypes';
 import muiTheme from '../constants/muiTheme';
+import ListTitlteForm from './ListTitlteForm';
 
 const listTilteWrapperStyles = {
   display: 'flex',
@@ -42,19 +37,8 @@ type ListTitleProps = {
   dispatch: AppDispatch;
 };
 
-type ListTitleField = {
-  title: string;
-};
-
 const ListTitle: React.FC<ListTitleProps> = ({ column, dispatch }) => {
   const [isEdit, setIsEdit] = useState<boolean>(false);
-  const { fields } = EDIT_COLUMN_TITLE;
-  const [field] = fields as FormField[];
-  const {
-    handleSubmit,
-    control,
-    formState: { errors },
-  } = useForm<ListTitleField>({ mode: 'onChange' });
   const { t } = useTranslation();
 
   const showInput = () => {
@@ -65,88 +49,36 @@ const ListTitle: React.FC<ListTitleProps> = ({ column, dispatch }) => {
     setIsEdit(false);
   };
 
-  const onSubmit = (data: ListTitleField) => {
-    hideInput();
-    dispatch(setColumn({ ...column, title: data.title }));
-    dispatch(editColumn()).then(() => {
-      dispatch(setColumn(DEFAULT_COLUMN));
-    });
-  };
-
   const deleteColumn = () => {
     dispatch(setColumn(column));
     dispatch(openModal(REMOVE_COLUMN));
   };
 
-  if (isEdit) {
-    return (
-      <>
-        <Box
-          component="form"
-          onSubmit={handleSubmit(onSubmit)}
-          id="column-title-form"
-          sx={{ bgcolor: 'white', borderRadius: '5px', p: '0.2rem 1rem' }}
-        >
-          <Controller
-            key={field.name}
-            name={field.name as keyof ListTitleField}
-            control={control}
-            rules={convertRulesRegExp(field.registerOptions)}
-            defaultValue={column.title}
-            render={({ field: { onChange, value } }) => (
-              <TextField
-                variant="standard"
-                type={field.type}
-                placeholder={t(field.placeholder)}
-                fullWidth
-                onChange={onChange}
-                value={value}
-                autoComplete="off"
-                error={!!errors[field.name as keyof typeof errors]}
-                helperText={
-                  errors[field.name as keyof typeof errors]
-                  && t(`${errors[field.name as keyof typeof errors]?.message}`)
-                }
-              />
-            )}
-          />
-        </Box>
-        <Tooltip title={t('modal.submit')}>
-          <IconButton type="submit" form="column-title-form">
-            <DoneSharpIcon color="success" />
-          </IconButton>
-        </Tooltip>
-        <Tooltip title={t('modal.close')}>
-          <IconButton onClick={hideInput}>
-            <CloseSharpIcon color="action" />
-          </IconButton>
-        </Tooltip>
-      </>
-    );
-  }
-
   return (
-    <Box sx={listTilteWrapperStyles}>
-      <Tooltip title={t('boardPage.editBtn')} placement="right">
-        <Typography
-          onClick={showInput}
-          variant="h6"
-          sx={listTitleStyle}
-        >
-          {column.title}
-        </Typography>
-      </Tooltip>
-      <MoreVertIcon
-        onClick={showInput}
-        color="action"
-        sx={{ ml: '0.5rem', mr: 'auto' }}
-      />
-      <Tooltip title={t('boardPage.deleteBtn')} placement="right" sx={{ justifySelf: 'flex-end' }}>
-        <IconButton onClick={deleteColumn}>
-          <DeleteIcon />
-        </IconButton>
-      </Tooltip>
-    </Box>
+    isEdit ? <ListTitlteForm hideInput={hideInput} dispatch={dispatch} column={column} />
+      : (
+        <Box sx={listTilteWrapperStyles}>
+          <Tooltip title={t('boardPage.editBtn')} placement="right">
+            <Typography
+              onClick={showInput}
+              variant="h6"
+              sx={listTitleStyle}
+            >
+              {column.title}
+            </Typography>
+          </Tooltip>
+          <MoreVertIcon
+            onClick={showInput}
+            color="action"
+            sx={{ ml: '0.5rem', mr: 'auto' }}
+          />
+          <Tooltip title={t('boardPage.deleteBtn')} placement="right" sx={{ justifySelf: 'flex-end' }}>
+            <IconButton onClick={deleteColumn}>
+              <DeleteIcon />
+            </IconButton>
+          </Tooltip>
+        </Box>
+      )
   );
 };
